Add a test that makes getting in lib.bup.vfs.Node.fullname assert.
As far as I can see it is necessary for a nested subdirectory to have
the same name as the top level directory containing it. That's hard to
arrange without knowing where the bup source directory lives, so
unlike the rest of test code I assume and make use of a writable
directory /tmp.

I don't have a fix yet so this change makes the test suite break.